[NFC] BridgeJS: Move CodeFragmentPrintetr to BridgeJSCore
So that we can use it for Swift code generation
