+++ Plugins/BridgeJS/Sources/BridgeJSLink/JSIntrinsicRegistry.swift
@@ -1,0 +1,37 @@
+#if canImport(BridgeJSUtilities)
+import BridgeJSUtilities
+#endif
+
+/// Registry for JS helper intrinsics used during code generation.
+final class JSIntrinsicRegistry {
+    private var entries: [String: [String]] = [:]
+
+    var isEmpty: Bool {
+        entries.isEmpty
+    }
+
+    func register(name: String, build: (CodeFragmentPrinter) throws -> Void) rethrows {
+        guard entries[name] == nil else { return }
+        let printer = CodeFragmentPrinter()
+        try build(printer)
+        entries[name] = printer.lines
+    }
+
+    func reset() {
+        entries.removeAll()
+    }
+
+    func emitLines() -> [String] {
+        var emitted: [String] = []
+        for key in entries.keys.sorted() {
+            if let lines = entries[key] {
+                emitted.append(contentsOf: lines)
+                emitted.append("")
+            }
+        }
+        if emitted.last == "" {
+            emitted.removeLast()
+        }
+        return emitted
+    }
+}
