+++ Plugins/BridgeJS/Sources/BridgeJSUtilities/Utilities.swift
@@ -27,3 +27,50 @@ public enum BridgeJSGeneratedFile {
         """
     }
 }
+
+/// A printer for code fragments.
+public final class CodeFragmentPrinter {
+    public private(set) var lines: [String] = []
+    private var indentLevel: Int = 0
+
+    public init(header: String = "") {
+        self.lines.append(contentsOf: header.split(separator: "\n").map { String($0) })
+    }
+
+    public func nextLine() {
+        lines.append("")
+    }
+
+    public func write<S: StringProtocol>(_ line: S) {
+        if line.isEmpty {
+            // Empty lines should not have trailing spaces
+            lines.append("")
+            return
+        }
+        lines.append(String(repeating: " ", count: indentLevel * 4) + String(line))
+    }
+
+    public func write(lines: [String]) {
+        for line in lines {
+            write(line)
+        }
+    }
+
+    public func write(contentsOf printer: CodeFragmentPrinter) {
+        self.write(lines: printer.lines)
+    }
+
+    public func indent() {
+        indentLevel += 1
+    }
+
+    public func unindent() {
+        indentLevel -= 1
+    }
+
+    public func indent(_ body: () throws -> Void) rethrows {
+        indentLevel += 1
+        try body()
+        indentLevel -= 1
+    }
+}
